Moved all classes in 'org.eclipse.smarthome.core.common' to 'org.eclipse.smarthome.core.common.osgi'. Added language file in binding archetype.
