+++ bundles/config/org.eclipse.smarthome.config.xml/META-INF/MANIFEST.MF
@@ -7,7 +7,7 @@ Bundle-Vendor: Eclipse.org/SmartHome
 Bundle-RequiredExecutionEnvironment: JavaSE-1.7
 Import-Package: org.eclipse.smarthome.config.core,
  org.eclipse.smarthome.config.core.i18n,
- org.eclipse.smarthome.core.common,
+ org.eclipse.smarthome.core.common.osgi,
  org.eclipse.smarthome.core.i18n,
  org.osgi.framework,
  org.osgi.util.tracker,
