+++ bundles/core/org.eclipse.smarthome.core.binding.xml/META-INF/MANIFEST.MF
@@ -13,7 +13,7 @@ Import-Package: com.thoughtworks.xstream,
  org.eclipse.smarthome.config.xml.osgi,
  org.eclipse.smarthome.config.xml.util,
  org.eclipse.smarthome.core.binding,
- org.eclipse.smarthome.core.common,
+ org.eclipse.smarthome.core.common.osgi,
  org.eclipse.smarthome.core.i18n,
  org.osgi.framework,
  org.slf4j
