+++ bundles/core/org.eclipse.smarthome.core.thing.xml/META-INF/MANIFEST.MF
@@ -12,7 +12,7 @@ Import-Package: com.thoughtworks.xstream,
  org.eclipse.smarthome.config.xml,
  org.eclipse.smarthome.config.xml.osgi,
  org.eclipse.smarthome.config.xml.util,
- org.eclipse.smarthome.core.common,
+ org.eclipse.smarthome.core.common.osgi,
  org.eclipse.smarthome.core.i18n,
  org.eclipse.smarthome.core.thing,
  org.eclipse.smarthome.core.thing.binding,
