+++ bundles/core/org.eclipse.smarthome.core/src/main/java/org/eclipse/smarthome/core/internal/i18n/ResourceBundleTracker.java
@@ -11,7 +11,7 @@ import java.util.Collection;
 import java.util.LinkedHashMap;
 import java.util.Map;
 
-import org.eclipse.smarthome.core.common.ResolvedBundleTracker;
+import org.eclipse.smarthome.core.common.osgi.ResolvedBundleTracker;
 import org.osgi.framework.Bundle;
 import org.osgi.framework.BundleContext;
