+++ tools/archetype/binding/src/main/resources/archetype-resources/ESH-INF/i18n/__bindingId___xx_XX.properties
@@ -1,0 +1,11 @@
+# binding
+binding.${bindingId}.name = <Your localized Binding name>
+binding.${bindingId}.description = <Your localized Binding description>
+
+# thing types
+thing-type.${bindingId}.sample.label = <Your localized Thing label>
+thing-type.${bindingId}.sample.description = <Your localized Thing description>
+
+# channel types
+channel-type.${bindingId}.sample-channel.label = <Your localized Channel label>
+channel-type.${bindingId}.sample-channel.description = <Your localized Channel description>
